requirements.txt
fastapi
uvicorn
chromadb
groq
python-dotenv
sentence-transformers
PyPDF2
numpy
pandas
